size long int

+++ src/utils.cpp
@@ -56,7 +56,7 @@ void msg(const string& text, const string& title) {
 }
 
 
-float *get_fmem(int size, char *str){
+float *get_fmem(long int size, char *str){
     // Careful with memory overcommitment:
     // https://stackoverflow.com/questions/48585079/malloc-on-linux-without-overcommitting
     // TODO: This function does not work properly (...but it does, at least most of the time -for linux and mac-)
